examplebroker: Move users to another file
We can then expose these cleanly
